parametrizacion de bancos
